try to handle proper cancellation
