+++ apm-agent-plugins/apm-reactor-plugin/src/test/java/co/elastic/apm/agent/reactor/TracedSubscriberTest.java
@@ -91,7 +91,7 @@ class TracedSubscriberTest extends AbstractInstrumentationTest {
         TracedSubscriber.unregisterHooks();
         TracedSubscriber.registerHooks(tracer);
 
-        flushGcExpiry(TracedSubscriber.getContextMap(), 1);
+        flushGcExpiry(TracedSubscriber.getContextMap(), 3);
     }
 
     @Test
